feat: restructure backend with new server and database configuration
- Updated entry point in package.json to src/server.js
- Created src/server.js to initialize Express server and set up middleware
- Moved database connection logic to src/config/db.js and added initialization function
- Implemented transaction management with CRUD operations in src/controllers/transactionController.js
- Added rate limiting middleware in src/middleware/rateLimiter.js
- Defined transaction routes in src/routes/transactionRouter.js
- Removed obsolete db.js configuration file
