+++ src/config/upstash.js
@@ -5,7 +5,7 @@ dotenv.config();
 
 const rateLimit = new Ratelimit({
   redis: Redis.fromEnv(),
-  limiter: Ratelimit.slidingWindow(10, "60 s"), // 10 requests per minute
+  limiter: Ratelimit.slidingWindow(100, "60 s"), // 100 requests per minute
 });
 
 module.exports = { rateLimit };
